requirements.txt
absl-py==0.9.0
anndata==0.7.3
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
bokeh==2.1.1
click==7.1.2
cloudpickle==1.4.1
colorcet==2.0.2
contextvars==2.4
cycler==0.10.0
Cython==0.29.21
dask==2.19.0
datashader==0.11.0
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
distributed==2.19.0
entrypoints==0.3
fsspec==0.7.4
gast==0.3.3
get-version==2.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hdbscan==0.8.26
HeapDict==1.0.1
holoviews==1.13.3
immutables==0.14
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
legacy-api-wrap==1.2
leidenalg==0.8.1
llvmlite==0.33.0
locket==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
msgpack==1.0.0
multipledispatch==0.6.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.18.2
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.7.0
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pyct==0.4.6
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2019.3
pyviz-comms==0.7.5
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
scanpy==1.5.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.1
Send2Trash==1.5.0
setuptools-scm==4.1.2
six==1.14.0
sortedcontainers==2.2.2
statsmodels==0.11.1
tables==3.6.1
tbb==2020.0.133
tblib==1.6.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typing-extensions==3.7.4.2
umap-learn==0.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.15.1
zict==2.0.0
zipp==1.2.0
